refactor(todo-list): convert model to entity
Move to its own file
Re-wire imports
